Extracted normalize fn.
