+++ src/duration/calculations.js
@@ -1,43 +1,6 @@
-import {
-  addMicroseconds,
-  addMilliseconds,
-  addSeconds,
-  addMinutes,
-  addHours,
-  addDays,
-  addWeeks,
-  addMonths,
-  addYears,
-} from './calculations/add';
-
-import {
-  ONE_HOUR,
-  ONE_MINUTE,
-  ONE_SECOND,
-} from './constants';
-
-import {
-  toIso,
-  toFragments,
-  asMicroseconds,
-  removeDateComponent,
-} from './index';
-
-export const normalize = (isoString) => {
-  const fragments = toFragments(isoString);
-
-  const microseconds = asMicroseconds(removeDateComponent(isoString));
-  const hours = Math.floor(microseconds / ONE_HOUR);
-  const minutes = Math.floor((microseconds - hours * ONE_HOUR) / ONE_MINUTE);
-  const seconds = ((microseconds - minutes * ONE_MINUTE) - (hours * ONE_HOUR)) / ONE_SECOND;
-
-  return toIso({
-    ...fragments,
-    hours,
-    minutes,
-    seconds,
-  });
-};
+export {
+  normalize
+} from './calculations/normalize';
 
 export {
   addMicroseconds,
